[SYCL][COMPAT] util header split in math and util headers (#12957)
This PR prepares SYCLcompat to receive multiple PRs containing updates
to the helper headers.

It will substitute the previous opened PR as it grew in scope too much,
as in case of issues it would be difficult to track and resolve
effectively the problem.

Edit: Previous PR was https://github.com/intel/llvm/pull/11267
