+++ sycl/include/syclcompat/util.hpp
@@ -31,14 +31,12 @@
 
 #pragma once
 
-#define SYCL_EXT_ONEAPI_COMPLEX
-
 #include <cassert>
-#include <complex>
-#include <sycl/ext/oneapi/experimental/complex/complex.hpp>
-#include <sycl/sycl.hpp>
 #include <type_traits>
 
+#include <sycl/sycl.hpp>
+
+#include <syclcompat/math.hpp>
 #include <syclcompat/memory.hpp>
 
 namespace syclcompat {
@@ -48,23 +46,23 @@ template <typename T> struct DataType {
   using T2 = T;
 };
 template <typename T> struct DataType<sycl::vec<T, 2>> {
-  using T2 = sycl::ext::oneapi::experimental::complex<T>;
+  using T2 = detail::complex_type<T>;
 };
 
 inline void matrix_mem_copy(void *to_ptr, const void *from_ptr, int to_ld,
                             int from_ld, int rows, int cols, int elem_size,
-                            sycl::queue queue = get_default_queue(),
+                            sycl::queue queue = syclcompat::get_default_queue(),
                             bool async = false) {
   if (to_ptr == from_ptr && to_ld == from_ld) {
     return;
   }
 
   if (to_ld == from_ld) {
-    size_t cpoy_size = elem_size * ((cols - 1) * to_ld + rows);
+    size_t copy_size = elem_size * ((cols - 1) * (size_t)to_ld + rows);
     if (async)
-      detail::memcpy(queue, (void *)to_ptr, (void *)from_ptr, cpoy_size);
+      detail::memcpy(queue, (void *)to_ptr, (void *)from_ptr, copy_size);
     else
-      detail::memcpy(queue, (void *)to_ptr, (void *)from_ptr, cpoy_size).wait();
+      detail::memcpy(queue, (void *)to_ptr, (void *)from_ptr, copy_size).wait();
   } else {
     if (async)
       detail::memcpy(queue, to_ptr, from_ptr, elem_size * to_ld,
@@ -117,96 +115,6 @@ inline double cast_ints_to_double(int high32, int low32) {
   sycl::int2 v0{high32, low32};
   auto v1 = v0.as<sycl::vec<double, 1>>();
   return v1;
-}
-
-/// Compute fast_length for variable-length array
-/// \param [in] a The array
-/// \param [in] len Length of the array
-/// \returns The computed fast_length
-inline float fast_length(const float *a, int len) {
-  switch (len) {
-  case 1:
-    return sycl::fast_length(a[0]);
-  case 2:
-    return sycl::fast_length(sycl::float2(a[0], a[1]));
-  case 3:
-    return sycl::fast_length(sycl::float3(a[0], a[1], a[2]));
-  case 4:
-    return sycl::fast_length(sycl::float4(a[0], a[1], a[2], a[3]));
-  case 0:
-    return 0;
-  default:
-    float f = 0;
-    for (int i = 0; i < len; ++i)
-      f += a[i] * a[i];
-    return sycl::sqrt(f);
-  }
-}
-
-/// Compute vectorized max for two values, with each value treated as a vector
-/// type \p S
-/// \param [in] S The type of the vector
-/// \param [in] T The type of the original values
-/// \param [in] a The first value
-/// \param [in] b The second value
-/// \returns The vectorized max of the two values
-template <typename S, typename T> inline T vectorized_max(T a, T b) {
-  sycl::vec<T, 1> v0{a}, v1{b};
-  auto v2 = v0.template as<S>();
-  auto v3 = v1.template as<S>();
-  v2 = sycl::max(v2, v3);
-  v0 = v2.template as<sycl::vec<T, 1>>();
-  return v0;
-}
-
-/// Compute vectorized min for two values, with each value treated as a vector
-/// type \p S
-/// \param [in] S The type of the vector
-/// \param [in] T The type of the original values
-/// \param [in] a The first value
-/// \param [in] b The second value
-/// \returns The vectorized min of the two values
-template <typename S, typename T> inline T vectorized_min(T a, T b) {
-  sycl::vec<T, 1> v0{a}, v1{b};
-  auto v2 = v0.template as<S>();
-  auto v3 = v1.template as<S>();
-  v2 = sycl::min(v2, v3);
-  v0 = v2.template as<sycl::vec<T, 1>>();
-  return v0;
-}
-
-/// Compute vectorized isgreater for two values, with each value treated as a
-/// vector type \p S
-/// \param [in] S The type of the vector
-/// \param [in] T The type of the original values
-/// \param [in] a The first value
-/// \param [in] b The second value
-/// \returns The vectorized greater than of the two values
-template <typename S, typename T> inline T vectorized_isgreater(T a, T b) {
-  sycl::vec<T, 1> v0{a}, v1{b};
-  auto v2 = v0.template as<S>();
-  auto v3 = v1.template as<S>();
-  auto v4 = sycl::isgreater(v2, v3);
-  v0 = v4.template as<sycl::vec<T, 1>>();
-  return v0;
-}
-
-/// Compute vectorized isgreater for two unsigned int values, with each value
-/// treated as a vector of two unsigned short
-/// \param [in] a The first value
-/// \param [in] b The second value
-/// \returns The vectorized greater than of the two values
-template <>
-inline unsigned vectorized_isgreater<sycl::ushort2, unsigned>(unsigned a,
-                                                              unsigned b) {
-  sycl::vec<unsigned, 1> v0{a}, v1{b};
-  auto v2 = v0.template as<sycl::ushort2>();
-  auto v3 = v1.template as<sycl::ushort2>();
-  sycl::ushort2 v4;
-  v4[0] = v2[0] > v3[0];
-  v4[1] = v2[1] > v3[1];
-  v0 = v4.template as<sycl::vec<unsigned, 1>>();
-  return v0;
 }
 
 /// Reverse the bit order of an unsigned integer
@@ -359,49 +267,6 @@ T permute_sub_group_by_xor(sycl::sub_group g, T x, unsigned int mask,
                                  target_offset < logical_sub_group_size
                                      ? start_index + target_offset
                                      : id);
-}
-
-/// Computes the multiplication of two complex numbers.
-/// \tparam T Complex element type
-/// \param [in] x The first input complex number
-/// \param [in] y The second input complex number
-/// \returns The result
-template <typename T>
-sycl::vec<T, 2> cmul(sycl::vec<T, 2> x, sycl::vec<T, 2> y) {
-  sycl::ext::oneapi::experimental::complex<T> t1(x[0], x[1]), t2(y[0], y[1]);
-  t1 = t1 * t2;
-  return sycl::vec<T, 2>(t1.real(), t1.imag());
-}
-
-/// Computes the division of two complex numbers.
-/// \tparam T Complex element type
-/// \param [in] x The first input complex number
-/// \param [in] y The second input complex number
-/// \returns The result
-template <typename T>
-sycl::vec<T, 2> cdiv(sycl::vec<T, 2> x, sycl::vec<T, 2> y) {
-  sycl::ext::oneapi::experimental::complex<T> t1(x[0], x[1]), t2(y[0], y[1]);
-  t1 = t1 / t2;
-  return sycl::vec<T, 2>(t1.real(), t1.imag());
-}
-
-/// Computes the magnitude of a complex number.
-/// \tparam T Complex element type
-/// \param [in] x The input complex number
-/// \returns The result
-template <typename T> T cabs(sycl::vec<T, 2> x) {
-  sycl::ext::oneapi::experimental::complex<T> t(x[0], x[1]);
-  return abs(t);
-}
-
-/// Computes the complex conjugate of a complex number.
-/// \tparam T Complex element type
-/// \param [in] x The input complex number
-/// \returns The result
-template <typename T> sycl::vec<T, 2> conj(sycl::vec<T, 2> x) {
-  sycl::ext::oneapi::experimental::complex<T> t(x[0], x[1]);
-  t = conj(t);
-  return sycl::vec<T, 2>(t.real(), t.imag());
 }
 
 /// Inherited from the original SYCLomatic compatibility headers.
